Fixes for PR comments
