+++ frontend/src/utils/api.ts
@@ -10,6 +10,6 @@ export const api = {
     CHANGE_PASSWORD: '/api/change_password/',
     VIEW_SCHEDULED_EVENTS: '/api/scheduled_event/',
     FETCH_RUN_INSTANCES: '/api/scheduled_event/{id}/run_instances/',
-    CREATE_SCHEDULED_INSTANCE: '/run_instances/',
+    CREATE_SCHEDULED_INSTANCE: '/api/scheduled_event/{scheduleId}/run_instances/',
   },
 };
